+++ crates/astria-sequencer-relayer/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update `idna` dependency to resolve cargo audit warning [#1869](https://github.com/astriaorg/astria/pull/1869).
+- Use `TxStatus` gRPC to confirm transaction commitment instead of `GetTx` [#1940](https://github.com/astriaorg/astria/pull/1940).
 
 ## [1.0.0] - 2024-10-25